implemented anonymousId cart functionality
now you are able to refresh the page without losing the cart content .
UUID, cookies js and commercetools Custom Objects to store cart data

still having timing issues when creating a new cart because of the cookies  when the create new cart button is clicked inconsistency, timing or synchronization issue.
The issue is that the anonymous id is created, then the cart is fetched, before the /api/anonymous-cart endpoint has finished creating the custom object.